dev-requirements.txt
ipython==8.11.0
pytest==7.2.2
django-extensions==3.2.1
